+++ tests/test_directives/test_directive_other.py
@@ -138,10 +138,10 @@ def test_reversed_toctree(app):
 
 @pytest.mark.sphinx(testroot='toctree-glob')
 def test_toctree_class(app):
-    text = (".. toctree::\n"
-            "   :class: custom-toc\n"
-            "\n"
-            "   foo\n")
+    text = ('.. toctree::\n'
+            '   :class: custom-toc\n'
+            '\n'
+            '   foo\n')
     app.env.find_files(app.config, app.builder)
     doctree = restructuredtext.parse(app, text, 'index')
     assert_node(doctree, [nodes.document, nodes.compound, addnodes.toctree])