+++ CHANGES.rst
@@ -12,8 +12,6 @@ Deprecated
 
 Features added
 --------------
-* #12524: Add ``class`` option to ``toctree`` directive.
-  Patch by Tim Hoffmann.
 * #11165: Support the `officially recommended`_ ``.jinja`` suffix for template
   files.
   Patch by James Addison and Adam Turner
@@ -57,6 +55,8 @@ Features added
 
 * #12258: Support ``typing_extensions.Unpack``
   Patch by Bénédikt Tran and Adam Turner.
+* #12524: Add ``class`` option to ``toctree`` directive.
+  Patch by Tim Hoffmann.
 
 Bugs fixed
 ----------